+++ src/app/app-routing.module.ts
@@ -1,19 +1,22 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { ClientComponent } from './pages/client/client.component';
-import { HomeComponent } from './pages/home/home.component';
 
 const routes: Routes = [
-  { path: 'home', component: HomeComponent },
-  { path: 'client', component: ClientComponent },
-  // { path: 'product', component: ProductComponent },
-  // {
-  //   path: 'product',
-  //   loadChildren: () =>
-  //     import('../app/pages/product/product.module').then(
-  //       (m) => m.ProductModule
-  //     ),
-  // },
+  {
+    path: 'home',
+    loadChildren: () =>
+      import('./pages/home/home.module').then((m) => m.HomeModule),
+  },
+  {
+    path: 'client',
+    loadChildren: () =>
+      import('./pages/client/client.module').then((m) => m.ClientModule),
+  },
+  {
+    path: 'product',
+    loadChildren: () =>
+      import('./pages/product/product.module').then((m) => m.ProductModule),
+  },
   { path: '', pathMatch: 'full', redirectTo: '/home' },
 ];
