+++ src/app/pages/client/client-routing.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ClientComponent } from './client.component';
+
+const routes: Routes = [
+  { path: '', component: ClientComponent}
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ClientRoutingModule { }
