+++ src/app/pages/client/client.module.ts
@@ -1,25 +1,14 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { SharedModule } from "../../shared/shared.module";
+
+import { HttpClientModule } from '@angular/common/http';
+import { SharedModule } from "../../shared/components/shared.module";
 import { ClientListComponent } from './client-list/client-list.component';
+import { ClientRoutingModule } from './client-routing.module';
 import { ClientComponent } from './client.component';
 
-
-
 @NgModule({
-    declarations: [
-        ClientComponent,
-        ClientListComponent
-    ],
-    exports: [
-        ClientComponent,
-        ClientListComponent
-    ],
-    imports: [
-        CommonModule,
-        HttpClientModule,
-        SharedModule
-    ]
+    declarations: [ClientComponent, ClientListComponent],
+    imports: [CommonModule, ClientRoutingModule, HttpClientModule, SharedModule]
 })
-export class ClientModule { }
+export class ClientModule {}
