+++ src/app/pages/home/home.component.html
@@ -5,5 +5,4 @@
   </div>
 
   <app-contact [socialLinks]="socialLinks"></app-contact>
-  <!-- <app-contact></app-contact> -->
 </section>
