+++ src/app/pages/home/home.module.ts
@@ -1,18 +1,19 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { ContactComponent } from './contact/contact.component';
 
+import { ContactComponent } from './contact/contact.component';
+import { HomeRoutingModule } from './home-routing.module';
+import { HomeComponent } from './home.component';
 
 
 @NgModule({
   declarations: [
+    HomeComponent,
     ContactComponent
   ],
   imports: [
-    CommonModule
-  ],
-  exports:[
-    ContactComponent
+    CommonModule,
+    HomeRoutingModule
   ]
 })
 export class HomeModule { }
