+++ src/app/pages/product/product-list/product-list.component.scss
@@ -11,9 +11,9 @@
 
   @media (max-width: 768px) {
     margin-left: 0;
-    padding-top: 100vh;
-    min-height: calc(100vh - 155px);
+    max-height: 700px;
     overflow-y: auto;
+    gap: 20px;
   }
 
   &__item {
