--- src/app/shared/shared-routing.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-@NgModule({
-  imports: [RouterModule],
-  exports: [RouterModule],
-})
-export class SharedRoutingModule {}
